ISIS-2177: rename OidFactory -> ObjectIdentifier, move closer to ObjectManager (responsibility: identify objects)
