+++ core/metamodel/src/main/java/org/apache/isis/metamodel/objectmanager/ObjectManager.java
@@ -21,6 +21,8 @@ package org.apache.isis.metamodel.objectmanager;
 import javax.annotation.Nullable;
 
 import org.apache.isis.metamodel.MetaModelContext;
+import org.apache.isis.metamodel.objectmanager.create.ObjectCreator;
+import org.apache.isis.metamodel.objectmanager.load.ObjectLoader;
 import org.apache.isis.metamodel.spec.ManagedObject;
 import org.apache.isis.metamodel.spec.ObjectSpecification;
