+++ core/metamodel/src/main/java/org/apache/isis/metamodel/objectmanager/identify/ObjectIdentifier_builtinHandlers.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.metamodel.adapter.oid.factory;
+package org.apache.isis.metamodel.objectmanager.identify;
 
 import java.util.UUID;
 
@@ -24,15 +24,15 @@ import org.apache.isis.commons.internal.exceptions._Exceptions;
 import org.apache.isis.config.SystemConstants;
 import org.apache.isis.metamodel.adapter.oid.Oid;
 import org.apache.isis.metamodel.adapter.oid.RootOid;
-import org.apache.isis.metamodel.adapter.oid.factory.OidFactory.Handler;
 import org.apache.isis.metamodel.facets.object.entity.EntityFacet;
 import org.apache.isis.metamodel.facets.object.value.ValueFacet;
 import org.apache.isis.metamodel.facets.object.viewmodel.ViewModelFacet;
+import org.apache.isis.metamodel.objectmanager.identify.ObjectIdentifier.Handler;
 import org.apache.isis.metamodel.spec.ManagedObject;
 
 import lombok.val;
 
-class OidFactory_builtinHandlers {
+class ObjectIdentifier_builtinHandlers {
 
 
     static class GuardAgainstRootOid implements Handler {
