+++ core/metamodel/src/main/java/org/apache/isis/metamodel/objectmanager/load/ObjectLoader.java
@@ -17,7 +17,7 @@
  *  under the License.
  */
 
-package org.apache.isis.metamodel.objectmanager;
+package org.apache.isis.metamodel.objectmanager.load;
 
 import org.apache.isis.commons.handler.ChainOfResponsibility;
 import org.apache.isis.commons.internal.collections._Lists;
